temporary editor permissions

+++ src/functions/api/endpoints/line/api.ts
@@ -8,19 +8,26 @@ import { validateLineGeoJson } from 'core/features/line/validations';
 import { nanoid } from 'nanoid';
 import { DDBLineDetailItem } from 'core/db/line/details/types';
 import * as turf from '@turf/turf';
-import { validateMapFeatureEditor } from 'core/features/mapFeature';
+import {
+  addTemporaryEditorToMapFeature,
+  validateMapFeatureEditor,
+  validateMapFeatureHasNoEditors,
+} from 'core/features/mapFeature';
 import { processLineGeoJson } from 'core/features/geojson';
 import { assignFromSourceToTarget } from 'core/utils';
 import { updateFeatureImagesInS3 } from 'core/features/mapFeature/image';
 import { logger } from 'core/utils/logger';
+import { accountApi } from 'core/externalApi/account-api';
 
 export const getLineDetails = async (req: Request, res: Response) => {
   const line = await db.getLineDetails(req.params.id);
   if (!line) {
     throw new Error(`NotFound: Line ${req.params.id} not found`);
   }
-  const isUserEditor = await validateMapFeatureEditor(line.lineId, req.user?.isaId);
-  res.json(getLineDetailsResponse(line, isUserEditor));
+  const isUserEditor = Boolean(await validateMapFeatureEditor(line.lineId, req.user?.isaId));
+  const hasNoEditors = await validateMapFeatureHasNoEditors(line.lineId);
+
+  res.json(getLineDetailsResponse(line, isUserEditor, hasNoEditors));
 };
 
 export const getLineGeoJson = async (req: Request, res: Response) => {
@@ -117,9 +124,21 @@ export const updateLine = async (req: Request<any, any, UpdateLinePostBody>, res
 
 export const deleteLine = async (req: Request, res: Response) => {
   const lineId = req.params.id;
-  await validateMapFeatureEditor(lineId, req.user?.isaId, true);
+  const editor = await validateMapFeatureEditor(lineId, req.user?.isaId, true);
+  if (editor?.grantedThrough === 'temporary') {
+    throw new Error('Forbidden: Cannot delete line with temporary editorship');
+  }
   await db.deleteLine(lineId);
   logger.info('deleted line', { user: req.user, lineId });
+  res.json({});
+};
+
+export const requestTemporaryEditorship = async (req: Request, res: Response) => {
+  const requestClaims = verifyRequestClaims(req);
+  const lineId = req.params.id;
+  await addTemporaryEditorToMapFeature(lineId, requestClaims.isaId);
+
+  logger.info('added temporary editor for line', { user: req.user, lineId });
   res.json({});
 };
 
@@ -129,3 +148,4 @@ lineApi.get('/:id/details', catchExpressJsErrorWrapper(getLineDetails));
 lineApi.get('/:id/geojson', catchExpressJsErrorWrapper(getLineGeoJson));
 lineApi.put('/:id', catchExpressJsErrorWrapper(updateLine));
 lineApi.delete('/:id', catchExpressJsErrorWrapper(deleteLine));
+lineApi.put('/:id/requestTemporaryEditorship', catchExpressJsErrorWrapper(requestTemporaryEditorship));
