+++ src/functions/api/endpoints/spot/api.ts
@@ -5,7 +5,7 @@ import { getSpotDetailsResponse } from './dto';
 import { FeatureCollection } from '@turf/turf';
 import { CreateSpotPostBody, createSpotSchema, UpdateSpotPostBody, updateSpotSchema } from './schema';
 import { processSpotGeoJson } from 'core/features/geojson';
-import { validateMapFeatureEditor } from 'core/features/mapFeature';
+import { addTemporaryEditorToMapFeature, validateMapFeatureEditor } from 'core/features/mapFeature';
 import { assignFromSourceToTarget } from 'core/utils';
 import { nanoid } from 'nanoid';
 import { validateSpotGeoJson } from 'core/features/spot/validations';
@@ -18,8 +18,10 @@ export const getSpotDetails = async (req: Request, res: Response) => {
   if (!spot) {
     throw new Error('NotFound: Spot not found');
   }
-  const isUserEditor = await validateMapFeatureEditor(spot.spotId, req.user?.isaId);
-  res.json(getSpotDetailsResponse(spot, isUserEditor));
+  const isUserEditor = Boolean(await validateMapFeatureEditor(spot.spotId, req.user?.isaId));
+  const hasNoEditors = (await db.getMapFeatureEditors(spot.spotId, { limit: 1 })).length === 0;
+
+  res.json(getSpotDetailsResponse(spot, isUserEditor, hasNoEditors));
 };
 
 export const getSpotGeoJson = async (req: Request, res: Response) => {
@@ -99,10 +101,22 @@ export const updateSpot = async (req: Request<any, any, UpdateSpotPostBody>, res
 
 export const deleteSpot = async (req: Request, res: Response) => {
   const spotId = req.params.id;
-  await validateMapFeatureEditor(spotId, req.user?.isaId, true);
+  const editor = await validateMapFeatureEditor(spotId, req.user?.isaId, true);
+  if (editor?.grantedThrough === 'temporary') {
+    throw new Error('Forbidden: Cannot delete line with temporary editorship');
+  }
   await db.deleteSpot(spotId);
 
   logger.info('deleted spot', { user: req.user, spotId });
+  res.json({});
+};
+
+export const requestTemporaryEditorship = async (req: Request, res: Response) => {
+  const requestClaims = verifyRequestClaims(req);
+  const spotId = req.params.id;
+  await addTemporaryEditorToMapFeature(spotId, requestClaims.isaId);
+
+  logger.info('added temporary editor for spot', { user: req.user, spotId });
   res.json({});
 };
 
@@ -112,3 +126,4 @@ spotApi.get('/:id/details', catchExpressJsErrorWrapper(getSpotDetails));
 spotApi.get('/:id/geojson', catchExpressJsErrorWrapper(getSpotGeoJson));
 spotApi.put('/:id', catchExpressJsErrorWrapper(updateSpot));
 spotApi.delete('/:id', catchExpressJsErrorWrapper(deleteSpot));
+spotApi.put('/:id/requestTemporaryEditorship', catchExpressJsErrorWrapper(requestTemporaryEditorship));
